split implementation out of index
