+++ packages/jellyfish-wallet/package.json
@@ -40,7 +40,10 @@
   "dependencies": {
     "@defichain/jellyfish-address": "0.0.0",
     "@defichain/jellyfish-crypto": "0.0.0",
-    "@defichain/jellyfish-transaction": "0.0.0",
+    "@defichain/jellyfish-network": "0.0.0",
+    "@defichain/jellyfish-transaction": "0.0.0"
+  },
+  "devDependencies": {
     "@defichain/jellyfish-transaction-signature": "0.0.0"
   }
 }
