+++ packages/jellyfish-wallet/tsconfig.build.json
@@ -10,6 +10,7 @@
   "references": [
     {"path": "../jellyfish-address/tsconfig.build.json"},
     {"path": "../jellyfish-crypto/tsconfig.build.json"},
+    {"path": "../jellyfish-network/tsconfig.build.json"},
     {"path": "../jellyfish-transaction/tsconfig.build.json"},
     {"path": "../jellyfish-transaction-signature/tsconfig.build.json"}
   ]
